Grandes melhorias e correções. Melhorado o texto do oeditar contrato html. Criado o item data_preenchimento para o form recibo, mais uma opção para o usuario. Criado um uuid basico para o usuario, para aumentar a segurança dos sarquivos e outros itens afins. DEFAULT_AUTO_FIELD adicionado no settings do projeto. Melhorias e correções no gerenciar parcelas do signals, as parcelas dos mêses iguais mantém o valor de seus codigos ao editar contrato. Corrigidos bugs da view recibos. Melhorias e correções da view tabela, agora cria pasta 'tabela_docs' se ela não existir.

+++ home/signals.py
@@ -1,7 +1,10 @@
+import os
 import random
 import string
 from dateutil.relativedelta import relativedelta
 
+from Alugue_seu_imovel import settings
+from django.contrib.auth.signals import user_logged_in, user_logged_out
 from django.db.models.signals import pre_delete, post_save, pre_save, post_delete
 from django.dispatch import receiver
 
@@ -12,9 +15,13 @@ from notifications.models import Notification
 
 
 def gerenciar_parcelas(instance_contrato):
-    # Editar as parcelas quando o contrato é editado:
-    Parcela.objects.filter(do_contrato=instance_contrato.pk).delete()
-    # RECRIA AS PARCELAS, MAS NÃO SERIA MELHOR MANTER RECIBOS TRUE PARA OS MESES QUE JÁ RECEBERAM RECIBO?
+    # Fazer algumas coisas se tiver parcelas anteriormente(acontece qdo o usuario modifica o período do contrato)
+    parcelas_anter = Parcela.objects.filter(do_contrato=instance_contrato).values()[:]
+
+    if parcelas_anter:
+        Parcela.objects.filter(do_contrato=instance_contrato.pk).delete()
+
+    # Criar as parcelas (acontece sempre (contrato criado e editado)):
     for x in range(0, instance_contrato.duracao):
         data_entrada = instance_contrato.data_entrada
         data = data_entrada.replace(day=instance_contrato.dia_vencimento) + relativedelta(months=x)
@@ -32,13 +39,27 @@ def gerenciar_parcelas(instance_contrato):
                 parcela.save()
                 break
 
-    # "Apagar" notificações de recibos que se referem a parcelas fora do novo período do
-    # contrato(data de entrada-saída)
-    notificacoes_user = Notification.objects.filter(recipient=instance_contrato.do_locador, deleted=False)
-    parcelas_pks = Parcela.objects.filter(do_contrato=instance_contrato).values_list('pk', flat=True)
-    for notificacao in notificacoes_user:
-        if notificacao.target not in parcelas_pks:
-            notificacao.delete()
+    if parcelas_anter:
+        # No ato de editar o período do contrato esta função irá copiar uns dados(recibo_entregue e codigo)
+        # das parcelas antigas(apagadas) que possuem a mesma data para as novas parcelas.
+        parcelas_novas = Parcela.objects.filter(do_contrato=instance_contrato)
+
+        datas_do_anterior_trat = []
+        for i in parcelas_anter:
+            datas_do_anterior_trat.append(str(i['data_pagm_ref']))
+
+        for parcela in parcelas_novas:
+            if str(parcela.data_pagm_ref) in datas_do_anterior_trat:
+                index = datas_do_anterior_trat.index(str(parcela.data_pagm_ref))
+                parcela.codigo = parcelas_anter[index]['codigo']
+                parcela.save(update_fields=['codigo'])
+
+        # "Apagar" notificações antigas deste contrato
+        notificacoes_user = Notification.objects.filter(recipient=instance_contrato.do_locador, deleted=False)
+        parcelas_pks = Parcela.objects.filter(do_contrato=instance_contrato).values_list('pk', flat=True)
+        for notificacao in notificacoes_user:
+            if notificacao.actor_object_id not in parcelas_pks:
+                notificacao.delete()
 
 
 def tratar_pagamentos(instance_contrato, delete=False):
@@ -80,22 +101,32 @@ def tratar_pagamentos(instance_contrato, delete=False):
                        f'parcela de {parcela.data_pagm_ref.strftime("%B/%Y").upper()} do contrato ' \
                        f'{parcela.do_contrato.codigo} foi detectado. Confirme a entrega do recibo.'
 
-            if parcela.tt_pago == valor_mensal and parcela.recibo_entregue is False and parcela.pk not in \
-                    lista_actor_object_id:
-                notify.send(sender=parcela, recipient=parcela.do_usuario, verb=f'Recibo',
-                            description=mensagem)
+            if parcela.tt_pago == valor_mensal and parcela.pk not in lista_actor_object_id:
+                notify.send(sender=parcela, recipient=parcela.do_usuario, verb=f'Recibo', description=mensagem)
+
+
+@receiver(user_logged_in)
+def usuario_fez_login(sender, user, **kwargs):
+    pass
+
+
+@receiver(user_logged_out)
+def usuario_fez_logout(sender, user, **kwargs):
+    file = rf'{settings.MEDIA_ROOT}/tabela_docs/tabela_{user.uuid}_{user}.pdf'
+    if os.path.exists(file):
+        os.remove(file)
 
 
 @receiver(pre_save, sender=Usuario)
 def usuario_save(sender, instance, **kwargs):
-    if instance.id is None:  # if Criado
+    if instance.pk is None:  # if Criado
         pass
     else:
         # Apaga todos os recibos em pdf do usuario(para que novos possam ser criados) quando se modifica informações
         # desta model contidas neles
-        ante = Usuario.objects.get(id=instance.pk)
+        ante = Usuario.objects.get(pk=instance.pk)
         if ante.RG != instance.RG or ante.CPF != instance.CPF or ante.first_name != instance.first_name or \
-                ante.last_name != instance.last_name:
+                ante.last_name != instance.last_name or ante.recibo_preenchimento != instance.recibo_preenchimento:
             contratos = Contrato.objects.filter(do_locador=instance)
             for contrato in contratos:
                 contrato.recibos_pdf.delete()
@@ -103,12 +134,12 @@ def usuario_save(sender, instance, **kwargs):
 
 @receiver(pre_save, sender=Locatario)
 def locatario_save(sender, instance, **kwargs):
-    if instance.id is None:  # if Criado
+    if instance.pk is None:  # if Criado
         pass
     else:
         # Apaga todos os recibos em pdf do locatario(para que novos possam ser criados) quando se modifica
         # informações desta model contidas neles
-        ante = Locatario.objects.get(id=instance.pk)
+        ante = Locatario.objects.get(pk=instance.pk)
         if ante.RG != instance.RG or ante.CPF != instance.CPF or ante.nome != instance.nome:
             contratos = Contrato.objects.filter(do_locatario=instance)
             for contrato in contratos:
@@ -117,7 +148,7 @@ def locatario_save(sender, instance, **kwargs):
 
 @receiver(pre_save, sender=Contrato)
 def contrato_save(sender, instance, **kwargs):
-    if instance.id is None:  # if Criado
+    if instance.pk is None:  # if Criado
         pass
     else:
         # Após modificar um contrato(parametros: duracao e data_entrada):
@@ -125,7 +156,7 @@ def contrato_save(sender, instance, **kwargs):
         # Recalc. as parcelas (model Parcela) pagas a partir do tt de pagamentos armazenados no seu respectivo contrato
         # Enviar as notificações relacionadas
         # com a função: tratar_pagamentos
-        ante = Contrato.objects.get(id=instance.pk)
+        ante = Contrato.objects.get(pk=instance.pk)
         if ante.duracao != instance.duracao or ante.data_entrada != instance.data_entrada:
             gerenciar_parcelas(instance_contrato=instance)
             tratar_pagamentos(instance_contrato=instance)
@@ -158,21 +189,7 @@ def contrato_update(sender, instance, created, **kwargs):
 
     if created:
         # Gera as parcelas quando o contrato é criado:
-        for x in range(0, contrato.duracao):
-            data_entrada = contrato.data_entrada
-            data = data_entrada.replace(day=contrato.dia_vencimento) + relativedelta(months=x)
-
-            codigos_existentes = list(
-                Parcela.objects.filter(do_contrato=contrato.pk).values("codigo").values_list('codigo', flat=True))
-            while True:
-                recibo_codigo = ''.join(
-                    random.choice(string.ascii_uppercase + string.ascii_lowercase + string.digits) for _ in
-                    range(6))
-                if recibo_codigo not in codigos_existentes:
-                    Parcela.objects.create(do_usuario=contrato.do_locador, do_contrato=contrato, do_imovel=imovel,
-                                           do_locatario=locatario, data_pagm_ref=data,
-                                           codigo=f'{recibo_codigo[:3]}-{recibo_codigo[3:]}')
-                    break
+        gerenciar_parcelas(contrato)
 
 
 @receiver(pre_delete, sender=Contrato)
